created server using http
node server is created by importing http module from Node and  is assing to the http object. The http object calls the createServer() method with a callback function, that takes request & response as a argument. Now the return object is stored in the const server and the server object is called with the method listen() to listen to the request on to the localhost:3005. The port number is assigned to any thousand port which is suitable. as usualy port starts from 80