Update app.js
In this request url is storred in const url and a if condition is applied to navigate to the '/form' page where a form is created with one input and a submit button. Sending the form input(s) via POST method define in the form tag as an attribute and action equals to '/formMessage'.